Change vessel speed in marketable windows to double

+++ com-mmxlabs-models-lng-transformer/com.mmxlabs.models.lng.transformer.ui/src/com/mmxlabs/models/lng/transformer/ui/analytics/MarketabilityCargoSequenceGenerator.java
@@ -100,7 +100,7 @@ public class MarketabilityCargoSequenceGenerator {
 	}
 
 	private @NonNull IModifiableSequence createDiversionSequence(final Schedule schedule, final @NonNull IResource targetResource, final ISequenceElement load, final ISequenceElement saleMarket,
-			final ISequence sequence, final SequencesAttributesProviderImpl providers, final Integer vesselSpeed) {
+			final ISequence sequence, final SequencesAttributesProviderImpl providers, final Double vesselSpeed) {
 
 		IVesselCharter vesselCharter = vesselProvider.getVesselCharter(targetResource);
 
@@ -245,7 +245,7 @@ public class MarketabilityCargoSequenceGenerator {
 
 	@NonNullByDefault
 	public void generateOptions(final Schedule schedule, final ISequences sequences, final List<@NonNull ISequenceElement> orderedCargoElements, final IResource targetResource,
-			final IPortSlot portSlot, final Integer vesselSpeed, final ToBooleanFunction<ISequences> action) {
+			final IPortSlot portSlot, final Double vesselSpeed, final ToBooleanFunction<ISequences> action) {
 
 		for (final ISequenceElement e : orderedCargoElements) {
 			if (!helper.checkResource(e, targetResource)) {
